+++ src/functions/atomic/example_bot_function.py
@@ -1,85 +1,64 @@
-"""Module implementation of the atomic function of the telegram bot. Example of implementation."""
+"""
+Модуль для получения случайных цитат из сериала Breaking Bad.
+"""
 
-import os
-import logging
 from typing import List
 import telebot
 from telebot import types
-from telebot.callback_data import CallbackData
-from bot_func_abc import AtomicBotFunctionABC
+import requests
+from base_atomic_bot_function import BaseAtomicBotFunction
 
-class AtomicExampleBotFunction(AtomicBotFunctionABC):
-    """Example of implementation of atomic function"""
 
-    commands: List[str] = ["example", "ebf"]
-    authors: List[str] = ["IHVH"]
-    about: str = "Пример функции бота!"
-    description: str = """В поле  *description* поместите подробную информацию о работе функции.
-    Описание способов использования, логики работы. Примеры вызова функции - /ebf 
-    Возможные параметры функции `/example`  """
-    state: bool = True
+class AtomicExampleBotFunction(BaseAtomicBotFunction):
+    """
+    Реализация атомарной функции для получения цитат Breaking Bad.
+    """
 
-    bot: telebot.TeleBot
-    example_keyboard_factory: CallbackData
+    commands: List[str] = ["quote"]
+    authors: List[str] = ["FeyBM"]
+    about: str = "Получение цитат из Breaking Bad!"
+    description: str = (
+        "Этот бот позволяет получать случайные цитаты из сериала Breaking Bad.\n"
+        "Для получения цитат используйте команду /quote <количество>."
+    )
 
     def set_handlers(self, bot: telebot.TeleBot):
-
-        self.bot = bot
-        self.example_keyboard_factory = CallbackData('t_key_button', prefix=self.commands[0])
+        """Устанавливает обработчики для команды."""
+        super().set_base_handlers(bot)
 
         @bot.message_handler(commands=self.commands)
-        def example_message_hendler(message: types.Message):
-            chat_id_msg = f"\nCHAT ID = {message.chat.id}"
-            msg = (
-                f"Ваш запрос обработан в AtomicExampleBotFunction! {chat_id_msg}\n"
-                f"USER ID = {message.from_user.id} \nEXAMPLETOKEN = {self.__get_example_token()}"
-            )
-            bot.send_message(text=msg, chat_id=message.chat.id, reply_markup=self.__gen_markup())
+        def send_quote(message: types.Message):
+            """Обрабатывает команду /quote."""
+            try:
+                num_quotes = int(message.text.split()[1])
+            except (IndexError, ValueError):
+                bot.send_message(
+                    message.chat.id,
+                    "Пожалуйста, укажите количество цитат. Пример: /quote 3"
+                )
+                return
 
-        @bot.callback_query_handler(func=None, config=self.example_keyboard_factory.filter())
-        def example_keyboard_callback(call: types.CallbackQuery):
-            callback_data: dict = self.example_keyboard_factory.parse(callback_data=call.data)
-            t_key_button = callback_data['t_key_button']
+            quotes = self._fetch_quotes(num_quotes)
+            if not quotes:
+                bot.send_message(message.chat.id, "Не удалось получить цитаты.")
+                return
 
-            match (t_key_button):
-                case ('cb_yes'):
-                    bot.answer_callback_query(call.id, "Ответ ДА!")
-                case ('cb_no'):
-                    bot.answer_callback_query(call.id, "Ответ НЕТ!")
-                case ('force_reply'):
-                    force_reply = types.ForceReply(selective=False)
-                    text = "Отправьте текст для обработки в process_next_step"
-                    bot.send_message(call.message.chat.id, text, reply_markup=force_reply)
-                    bot.register_next_step_handler(call.message, self.__process_next_step)
-                case _:
-                    bot.answer_callback_query(call.id, call.data)
+            for quote in quotes:
+                bot.send_message(message.chat.id, quote)
 
-    def __get_example_token(self):
-        token = os.environ.get("EXAMPLETOKEN")
-        return token
-
-    def __gen_markup(self):
-        markup = types.InlineKeyboardMarkup()
-        markup.row_width = 2
-        yes_callback_data = self.example_keyboard_factory.new(t_key_button="cb_yes")
-        no_callback_data = self.example_keyboard_factory.new(t_key_button="cb_no")
-        force_reply_callback_data = self.example_keyboard_factory.new(t_key_button="force_reply")
-        markup.add(
-            types.InlineKeyboardButton("Да", callback_data=yes_callback_data),
-            types.InlineKeyboardButton("Нет", callback_data=no_callback_data),
-            types.InlineKeyboardButton("ForceReply", callback_data=force_reply_callback_data)
-        )
-        return markup
-
-    def __process_next_step(self, message):
-        try:
-            chat_id = message.chat.id
-            txt = message.text
-            if txt != "exit":
-                force_reply = types.ForceReply(selective=False)
-                text = f"text = {txt}; chat.id = {chat_id}; \n Для выхода из диалога введите - exit"
-                msg = self.bot.send_message(message.chat.id, text, reply_markup=force_reply)
-                self.bot.register_next_step_handler(msg, self.__process_next_step)
-        except ValueError as ex:
-            logging.exception(ex)
-            self.bot.reply_to(message, f"Exception - {ex}")
+    def _fetch_quotes(self, num: int):
+        """
+        Получение случайных цитат из API Breaking Bad.
+        """
+        quotes = []
+        for _ in range(num):
+            try:
+                response = requests.get("https://api.breakingbadquotes.xyz/v1/quotes", timeout=5)
+                if response.status_code == 200:
+                    data = response.json()[0]
+                    quotes.append(f"Цитата: {data['quote']}\nАвтор: {data['author']}")
+                else:
+                    return []
+            except requests.RequestException:
+                return []
+        return quotes
